### 2025-08-24 c51e04f
docs(joss): finalize paper.md and paper.bib - Updated Statement of Need (clarified messaging layer vs DB clustering) - Synced paper.bib (11 unique references, all cited)
