+++ paper.md
@@ -13,76 +13,81 @@ authors:
 affiliations:
   - name: Independent Researcher
     index: 1
-date: 2025-08-23
+date: 2025-08-24
 bibliography: paper.bib
 ---
 
 # Summary
 
-`spring-batch-db-cluster-partitioning` is an open-source library that adds **database-driven clustering** to Spring Batch [@springbatch].
-It provides **node heartbeats, explicit partition lifecycle, and automatic failover** using only a shared relational database.
-This enables **horizontal scale** and **fault tolerance** for partitioned steps across multiple JVMs or machines **without external messaging middleware**, while remaining idiomatic to Spring Batch.
+`spring-batch-db-cluster-partitioning` is an open-source extension for Spring Batch [@springbatch].
+It introduces **database-driven clustering** that provides **node heartbeats, partition lifecycle tracking, and automatic failover** using a shared relational database.
+The framework enables **horizontal scale-out and fault tolerance** for partitioned jobs while remaining fully compatible with the Spring Batch programming model.
 
 # Statement of Need
 
-Spring Batch offers local parallelism within a single JVM [@springbatch_scaling] and remote partitioning that depends on messaging middleware [@spring_integration_remote_partitioning].
-Neither option supplies a **DB-first** coordination model with **node liveness tracking** and **transactional partition state** suitable for simple, predictable operations.
-Many Spring Batch deployments already include a relational database; leveraging it as the cluster’s coordination plane removes extra moving parts while improving visibility and recoverability.
-This library fills that gap: it integrates **heartbeats, partition assignment, status tracking, and safe re-assignment** directly into Spring Batch workflows.
+Spring Batch supports local parallelism [@springbatch_scaling] and remote partitioning through a messaging layer [@spring_integration_remote_partitioning].
+While effective, these modes require either execution within a single JVM or the addition of messaging infrastructure for coordination.  
+This framework instead uses the relational database itself as the coordination plane.  
+Nodes register with heartbeats, so the set of available workers is explicit and continuously visible, simplifying operations while improving **reproducibility, observability, and resilience**.
+
+Typical use cases include **ETL (Extract–Transform–Load) workflows** where large datasets must be cleaned, transformed, or aggregated in parallel.
+By coordinating work through the database, this framework allows such pipelines to be executed across multiple machines in a cluster without requiring brokers or external schedulers.
 
 # Functionality
 
-The main capabilities of the library include:
+The framework provides:
 
-- **Cluster node registry & heartbeats** in a `BATCH_NODES` table with configurable timeouts.
-- **Partition lifecycle management** tracked in `BATCH_PARTITIONS` (PENDING/CLAIMED/COMPLETED/FAILED).
-- **Automatic failover**: incomplete, `is_transferable` partitions are safely re-assigned on node failure.
-- **Master-per-job** design: the launcher of a job acts as its master; no global controller/election.
-- **Flexible distribution strategies**: round-robin, fixed-node, and dynamic.
-- **Database compatibility**: Oracle, PostgreSQL, MySQL, H2 (for testing).
-- **Spring-native integration**: jobs/steps defined using Spring Batch APIs remain unchanged.
+- **Cluster node registry**: active nodes register in a `BATCH_NODES` table with periodic heartbeats.
+- **Two-phase liveness detection**: nodes are first marked `UNREACHABLE`, then removed after a cleanup threshold, reducing false positives.
[…]
