+++ paper.bib
@@ -1,60 +1,86 @@
 @misc{springbatch,
-  title = {Spring Batch Project},
+  title   = {Spring Batch Project},
   howpublished = {\url{https://spring.io/projects/spring-batch}},
-  year = {2025}
+  year    = {2025}
 }
 
 @misc{springbatch_scaling,
-  title = {Spring Batch: Scaling and Parallel Processing},
+  title   = {Spring Batch: Scaling and Parallel Processing},
   howpublished = {\url{https://docs.spring.io/spring-batch/reference/scalability.html}},
-  year = {2025}
+  year    = {2025}
 }
 
 @misc{spring_integration_remote_partitioning,
-  title = {Remote Partitioning with Spring Batch and Spring Integration},
+  title   = {Remote Partitioning with Spring Batch and Spring Integration},
   howpublished = {\url{https://docs.spring.io/spring-batch/reference/partitioning.html}},
-  year = {2025}
+  year    = {2025}
 }
 
 @misc{scdf,
-  title = {Spring Cloud Data Flow},
+  title   = {Spring Cloud Data Flow},
   howpublished = {\url{https://dataflow.spring.io/}},
-  year = {2025}
-}
-
-@misc{k8s_jobs,
-  title = {Kubernetes},
-  howpublished = {\url{https://kubernetes.io/}},
-  year = {2025}
+  year    = {2025}
 }
 
 @misc{project_repo,
-  author = {Chejarla, Janardhan Reddy},
-  title  = {spring-batch-db-cluster-partitioning},
+  author  = {Chejarla, Janardhan Reddy},
+  title   = {spring-batch-db-cluster-partitioning},
   howpublished = {\url{https://github.com/jchejarla/spring-batch-db-cluster-partitioning}},
-  year = {2025}
+  year    = {2025}
 }
 
-@article{dean2008mapreduce,
-  title={MapReduce: simplified data processing on large clusters},
-  author={Dean, Jeffrey and Ghemawat, Sanjay},
-  journal={Communications of the ACM},
-  volume={51},
-  number={1},
-  pages={107--113},
-  year={2008}
+@article{deelman2015pegasus,
+  title   = {Pegasus: A Workflow Management System for Science Automation},
+  author  = {Deelman, Ewa and Vahi, Karan and Juve, Gideon and Rynge, Mats and Callaghan, Scott and Maechling, Philip J. and Mayani, Rajiv and Chen, Weiwei and Ferreira da Silva, Rafael and Livny, Miron and Wenger, Kent},
+  journal = {Future Generation Computer Systems},
+  volume  = {46},
+  pages   = {17--35},
+  year    = {2015},
+  doi     = {10.1016/j.future.2014.10.008}
 }
 
-@inproceedings{zaharia2012spark,
-  title={Resilient distributed datasets: A fault-tolerant abstraction for in-memory cluster computing},
-  author={Zaharia, Matei and Chowdhury, Mosharaf and Das, Tathagata and others},
-  booktitle={Proceedings of the 9th USENIX Symposium on Networked Systems Design and Implementation (NSDI)},
-  year={2012}
+@article{ludascher2006kepler,
+  title   = {Scientific Workflow Management and the Kepler System},
+  author  = {Lud{\"a}scher, Bertram and Altintas, Ilkay and Berkley, Chad and Higgins, Dan and Jaeger, Efrat and Jones, Matthew and Lee, Edward A. and Tao, Jing and Zhao, Yang},
+  journal = {Concurrency and Computation: Practice and Experience},
+  volume  = {18},
+  number  = {10},
+  pages   = {1039--1065},
+  year    = {2006},
[…]
