### 2025-08-24 #8856 @jchejarla
This library extends Spring Batch with a database-backed clustering layer that adds node heartbeats, partition lifecycle tracking, and automatic failover. It qualifies as research software because it enables data-intensive workflows to be executed reproducibly and fault-tolerantly across clusters, with execution state durably stored in a relational database to support repeatability and auditing. Typical research use cases include ETL pipelines in bioinformatics, social sciences, sensor data analysis, and digital humanities, where large datasets must be processed in parallel with reliable recovery. By staying entirely within the Spring ecosystem, the framework lowers the barrier for small and medium research groups to scale their workflows without additional messaging infrastructure. Peer-reviewed references such as Pegasus, Kepler, and Taverna are already cited with DOIs; for other systems papers, canonical URLs are provided where DOIs are not available.
