### 2025-09-27 #8856 @majensen
Hello @r3w0p - would you be willing to review [this work](#8856) for [JOSS](https://joss.theoj.org)? We could use your expertise - thanks for considering! --@majensen
