### 2025-09-27 #8856 @majensen
Hi @otiai10 -- do you have time to review [this work](#8856) for [JOSS](https://joss.theoj.org)? Thanks very much for considering - --@majensen
