### 2025-11-13 #8856 @danielskatz
I'm using https://fediscience.org/@danielskatz/115543348939515342 and https://bsky.app/profile/danielskatz.org/post/3m5jlywg57c2l to try to help. I'm also posting to a couple of RSE slacks.
