### 2025-11-18 #9460 @editorialbot
[REVIEW]: spring-batch-db-cluster-partitioning: Database-driven clustering with heartbeats and failover for Spring Batch
**Submitting author:** @jchejarla (<a href="http://orcid.org/0009-0002-4876-802X">Janardhan Reddy Chejarla</a>)
**Repository:** https://github.com/jchejarla/spring-batch-db-cluster-partitioning
**Branch with paper.md** (empty if default branch): main
**Version:** v2.0.0
