### 2026-01-21 #9460 @danielskatz
Hi - As the track editor, I wanted to check on how this is going, as I don't see that anything has changed in the last two weeks. @majensen - Can you update me on the overall status here and the next steps?
