### 2026-01-26 jchejarla/spring-batch-db-cluster-partitioning#5 @david-guzman
[JOSS review] Add README to examples
Please add a README to `examples/` explaining the use of the example application to showcase the functional claims of this Spring Batch extension

**Optional**: It would be helpful to use data from any scientific domain (bioinformatics, physics, chemistry, etc) rather than the customer data currently used in the repository

Ref https://github.com/openjournals/joss-reviews/issues/9460
