### 2026-01-26 jchejarla/spring-batch-db-cluster-partitioning#6 @david-guzman
[JOSS review] Expand Summary section in paper
> Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

It would help to rephrase and expand the Summary section to target a diverse, non-specialist audience. Ideally explaining further the use of messaging infrastructure vs relational database for coordination.

Ref https://github.com/openjournals/joss-reviews/issues/9460
