### 2026-01-26 jchejarla/spring-batch-db-cluster-partitioning#7 @david-guzman
[JOSS review] Add context in Statement of Need
>  A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

While it is explained in the paper that the requirements of Spring Batch may present challenges that this extension attempts to address. It would help the reader to present this within the context of research software. What are the domains Spring-based workflows are used? Who is the target audience?

Optional: Why a Spring-centric approach vs a more generic Jakarta EE one?

Ref: https://github.com/openjournals/joss-reviews/issues/9460
