### 2026-01-26 jchejarla/spring-batch-db-cluster-partitioning#8 @david-guzman
[JOSS review] Writing style and minor changes in paper
> Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

- [ ] The functionality, architecture, failure handling, configuration, and limitation sections should be rewritten in prose, ideally justifying decisions on features supported, tasks and problems to address, etc. The style currently used, mostly lists, is more suitable for the README page in the repository. The paper would benefit by using lists only when prose cannot be used.
- [ ] The configuration section could be rewritten to explain how this extension can be used in practice. Explain the steps outlined in the Getting Started section of the README and reference the repository for details. The example content of Spring Boot properties might not be needed in the paper.
- [ ] Minor change: This is named "extension" and later "framework" in the paper. Best to use one for consistency.
- [ ] Minor change: In Limitations, "small-to-medium clusters". Any estimation on the number of nodes, etc.
- [ ] Minor change: Diagram is displaying the source code, diagram is not rendered

Ref https://github.com/openjournals/joss-reviews/issues/9460
