### 2026-01-26 #9460 @david-guzman
> [@Hardik27](https://github.com/Hardik27), [@david-guzman](https://github.com/david-guzman) have you been able to make progress reviewing this work? Let me know. Thanks

Apologies for the delay. I have created the following issues in the target repository

- [ ] https://github.com/jchejarla/spring-batch-db-cluster-partitioning/issues/5
- [ ] https://github.com/jchejarla/spring-batch-db-cluster-partitioning/issues/6
- [ ] https://github.com/jchejarla/spring-batch-db-cluster-partitioning/issues/7
- [ ] https://github.com/jchejarla/spring-batch-db-cluster-partitioning/issues/8
