### 2026-02-20 jchejarla/spring-batch-db-cluster-partitioning#9 @Hardik27
[JOSS review] Missing Required Section
The paper is missing a few required sections:

1. Summary
2. Software Design
3. Research Impact Statement
4. AI Usage Disclosure

Furthermore, upon closely reviewing the paper.md, the quality of writing and reference also needs a lot of work. For instance, the references are not in the citation syntax that are required.
