### 2026-02-20 jchejarla/spring-batch-db-cluster-partitioning#10 @Hardik27
[JOSS review] Test enhancement & Contributing guidelines.
The repository needs an automated test pipeline. Furthermore the contributing.md does not really mention any detail steps for other to contribute/report an issue. This is a must for any open source repository.
