### 2026-02-20 jchejarla/spring-batch-db-cluster-partitioning#11 @Hardik27
[JOSS review] Open-source Practoce
Based on the JOSS review guidelines, the software appears to have had a relatively limited development history in its current form. While the repository has received some attention (e.g., stars), there is little evidence of external contributions or community engagement. The authors may want to provide additional context on the maturity of the project, its usage, or plans for broader adoption and maintenance.
