### 2026-02-20 jchejarla/spring-batch-db-cluster-partitioning#12 @Hardik27
[JOSS review] Overall Paper.md modifications
Agreed with the author on the statement of need, however one thing that needs to be added is the benefit (in performance or any area) that the system gains upon partitioning. This would highly corroborate the statement of need. Furthermore, it is my professional opinion to highly modify the paper.md to go in more details for users to understand the technical innovation.
