### 2026-03-25 09f0475
Addressing the issues reported by reviewers during JOSS paper review
