+++ paper.bib
@@ -61,6 +61,15 @@
   doi     = {10.1093/nar/gkt328}
 }
 
+@book{tanenbaum2007distributed,
+  title   = {Distributed Systems: Principles and Paradigms},
+  author  = {Tanenbaum, Andrew S. and Van Steen, Maarten},
+  edition = {2nd},
+  publisher = {Pearson Prentice Hall},
+  year    = {2007},
+  isbn    = {978-0132392273}
+}
+
 @inproceedings{hunt2010zookeeper,
   title   = {ZooKeeper: Wait-free Coordination for Internet-scale Systems},
   author  = {Hunt, Patrick and Konar, Mahadev and Junqueira, Flavio P. and Reed, Benjamin},
