+++ paper.md
@@ -19,94 +19,61 @@ bibliography: paper.bib
 
 # Summary
 
-`spring-batch-db-cluster-partitioning` is an open-source extension for Spring Batch [@springbatch].
-It introduces **database-driven clustering** that provides **node heartbeats, partition lifecycle tracking, and automatic failover** using a shared relational database.
-The framework enables **horizontal scale-out and fault tolerance** for partitioned jobs while remaining fully compatible with the Spring Batch programming model.
+Batch processing of large datasets — transforming, validating, or migrating millions of records in a single run — is a routine requirement in domains ranging from financial services and healthcare to genomics and climate modelling. Spring Batch [@springbatch] is the most widely adopted Java framework for these workloads, providing transactional chunk processing, restart semantics, and a rich programming model. When a single machine cannot process the data fast enough, Spring Batch supports *remote partitioning*: the work is split into partitions and dispatched to worker nodes through a messaging layer such as RabbitMQ or Apache Kafka [@spring_integration_remote_partitioning].
+
+However, operating a highly available message broker solely for partition coordination adds significant infrastructure overhead — deployment, monitoring, and failure handling for a component that is incidental to the batch workload itself. More critically, the message-based model provides no built-in mechanism for the master node to discover how many workers are actually available, or to detect and recover from a worker crash after a partition has been dispatched.
+
+`spring-batch-db-cluster-partitioning` is an open-source Java extension that replaces the messaging layer with the relational database that Spring Batch already requires. It introduces three capabilities absent from the standard remote partitioning model: (1) **proactive node awareness**, allowing the master to query exactly how many workers are active before distributing work; (2) **explicit partition lifecycle tracking**, recording each partition's state (`PENDING`, `CLAIMED`, `COMPLETED`, `FAILED`) transactionally so that no work is silently lost; and (3) **automatic failover**, detecting unresponsive nodes via heartbeat timeouts and reassigning their incomplete partitions to healthy workers without human intervention. By persisting all coordination state in standard SQL tables, the extension eliminates the operational burden of a message broker while simultaneously providing real-time visibility into job progress through ordinary SQL queries.
 
 # Statement of Need
 
-Spring Batch supports local parallelism [@springbatch_scaling] and remote partitioning through a messaging layer [@spring_integration_remote_partitioning].
-While effective, these modes require either execution within a single JVM or the addition of messaging infrastructure for coordination.  
-This framework instead uses the relational database itself as the coordination plane.  
-Nodes register with heartbeats, so the set of available workers is explicit and continuously visible, simplifying operations while improving **reproducibility, observability, and resilience**.
+Spring Batch is widely used in research-adjacent and enterprise contexts where reproducible, auditable processing of large datasets is essential. In bioinformatics, Spring-based pipelines process sequencing data and clinical datasets at scale [@wolstencroft2013taverna]; in financial services, nightly batch runs reconcile millions of transactions; in climate and earth sciences, ETL workflows ingest observational data into analytical databases. In each case, horizontal scale-out — distributing partitions across multiple worker nodes — is a common strategy for meeting processing-time targets.
 
-Typical use cases include **ETL (Extract–Transform–Load) workflows** where large datasets must be cleaned, transformed, or aggregated in parallel.
-By coordinating work through the database, this framework allows such pipelines to be executed across multiple machines in a cluster without requiring brokers or external schedulers.
[…]
