+++ examples/README.md
@@ -1,0 +1,206 @@
+# Examples: Spring Batch DB Cluster Partitioning
+
+This folder contains a runnable example application that demonstrates the core functional claims of the
+**spring-batch-db-cluster-partitioning** extension:
+
+1. **Database-driven coordination** of partitioned Spring Batch steps across multiple workers/nodes.
+2. **Cluster-safe partition execution** with deterministic partition assignment and completion tracking.
+3. **Restartability and failure recovery** without reprocessing completed partitions.
+4. **Horizontal scalability**: increase throughput by adding workers without changing job logic.
+
+The example is intentionally minimal so that users can run it locally and validate behavior using only a database and a JVM.
+
+---
+
+## Example Jobs
+
+The application exposes two example jobs via REST endpoints:
+
+| Job | Endpoint | Description |
+|-----|----------|-------------|
+| **Range-sum (simple)** | `GET /api/v1/clusteredjob/addition/taskSize/{n}/from/{from}/to/{to}` | Splits a numeric range into `n` partitions, computes the sum in each partition on a different worker, and aggregates the result. |
+| **ETL (advanced)** | `GET /api/v1/etljob/rows/{rows}` | Reads rows from a CSV file (`data/customers_example_data.csv`), distributes chunks across workers using round-robin, and writes XML output. Demonstrates chunk-oriented reader/processor/writer with cluster partitioning. |
+
+There is also a **single-node baseline** endpoint (`GET /api/v1/singlenodejob/addition/...`) so you can compare execution time with and without clustering.
+
+---
+
+## Prerequisites
+
+- Java 17+ (or the version required by this repository)
+- Maven 3.x
+- A relational database supported by Spring Batch (PostgreSQL, MySQL, Oracle, or H2 for local testing)
+
+> **Tip:** Use PostgreSQL for the closest-to-production behavior. H2 is fine for quick local validation.
+
+---
+
+## Quick Start (single machine, multiple workers)
+
+### 1. Start the database
+
+If you use Docker:
+
+```bash
+docker run --name sb-db -e POSTGRES_PASSWORD=postgres -e POSTGRES_USER=postgres -e POSTGRES_DB=sb \
+  -p 5432:5432 -d postgres:16
+```
+
+Create the Spring Batch schema and the three cluster coordination tables. DDL scripts are provided in
+`spring-batch-db-cluster-core/src/main/resources/` (choose the file matching your database).
+
+### 2. Configure the database connection
+
+Edit `src/main/resources/application-postgres.yml` (or the profile matching your database) to point to your running instance:
+
+```yaml
+spring:
+  datasource:
+    url: jdbc:postgresql://localhost:5432/sb
+    username: postgres
+    password: postgres
+```
+
+The active profile is set in `application.properties`:
+
+```properties
+spring.profiles.active=postgres
+```
+
+Pre-configured profiles are available for **PostgreSQL**, **MySQL**, and **Oracle**.
+
+### 3. Build the project
+
+From the repository root:
+
+```bash
+mvn clean package -pl examples -am
+```
+
+### 4. Start Worker Node 1 (becomes master)
+
+```bash
+java -jar examples/target/examples-2.0.0.jar \
+  --server.port=8081 \
+  --spring.batch.cluster.node-id=worker-1
+```
+
+### 5. Start Worker Node 2
+
+In a separate terminal:
+
+```bash
+java -jar examples/target/examples-2.0.0.jar \
+  --server.port=8082 \
+  --spring.batch.cluster.node-id=worker-2
+```
+
+You can start additional workers on different ports in the same way.
+
+### 6. Trigger the job
+
+Send a request to any node. The node that receives the request becomes the master for that job execution:
+
+```bash
+# Range-sum job: 10 partitions, sum numbers from 1 to 1,000,000
[…]

[…]
