### 2026-03-25 jchejarla/spring-batch-db-cluster-partitioning#5 @jchejarla
Thank you for the suggestion. A comprehensive examples/README.md has been added covering:

1) Description of the two example jobs (range-sum and ETL) with their REST endpoints
2) Prerequisites and quick-start guide with Docker-based database setup
3) Step-by-step instructions to build, start multiple worker nodes, and trigger jobs
4) A dedicated Testing Failover section demonstrating how to validate partition reassignment
5) Monitoring via Actuator endpoints
6) Full project structure and configuration reference table
Regarding the optional suggestion to use scientific-domain data — the current example uses customer CSV data to keep things accessible, but I'm open to adding a second example with scientific data in a future iteration.
