### 2026-03-25 jchejarla/spring-batch-db-cluster-partitioning#6 @jchejarla
Thank you — the Summary has been rewritten and expanded to target a diverse, non specialist audience. It now:

1) Opens with a plain-English explanation of batch processing and the domains where it arises
2) explains why messaging infrastructure introduces operational overhead
3) Describes the three key capabilities this extension provides (proactive node awareness, partition lifecycle tracking, automatic failover)
4) Contrasts the messaging-based approach with the database-based coordination model
