### 2026-03-25 jchejarla/spring-batch-db-cluster-partitioning#7 @jchejarla
Addressed. The Statement of Need now includes:

**Research domains:** bioinformatics (sequencing/clinical datasets), financial services (transaction reconciliation), climate/earth sciences (observational data ETL)
**Target audience:** Java development teams needing horizontal scale-out or fault-tolerant execution, especially where a message broker is undesirable
**Why Spring-centric vs Jakarta EE:** Spring Batch is the dominant batch framework in the Java ecosystem; Jakarta EE lacks an equivalent batch-partitioning abstraction with comparable adoption — so a native Spring integration was the pragmatic choice
