### 2026-03-25 jchejarla/spring-batch-db-cluster-partitioning#8 @jchejarla
All points addressed:

**Prose style:** The Functionality, Architecture, Failure Handling, Configuration, and Limitations sections have been completely rewritten in continuous prose. No bullet-point lists remain in the paper.
**Configuration section:** The raw YAML block has been removed. Practical usage is now described in the Software Design section in prose, with the repository referenced for step-by-step Getting Started details.
**Consistent naming:** The project is now consistently referred to as "extension" throughout. "Framework" appears only when referring to Spring Batch or Spring Boot themselves.
**"Small-to-medium clusters" quantified:** Now reads "typically two to approximately twenty nodes."
**Diagram rendering:** Replaced the mermaid code block with a rendered PNG image reference (![...](docs/architecture.png)) that will display correctly in the JOSS PDF build.
