### 2026-03-25 jchejarla/spring-batch-db-cluster-partitioning#9 @jchejarla
Thank you for the detailed feedback. The paper has been substantially revised:

**Summary:** Expanded from 1 paragraph to 3 paragraphs targeting a non-specialist audience.
**Software Design:** New section added — describes the three coordination tables, the partitioning flow (master discovery -> partition assignment -> worker polling -> claim-> execution), and the two-phase failure detection protocol. Includes an architecture diagram.
**Research Impact:** New standalone section explaining how the extension benefits research teams by lowering infrastructure barriers and supporting reproducibility through auditable SQL state.
**AI Usage Disclosure:** New section added.
**References:** All citations now use proper pandoc [@key] syntax, and every citation key has a corresponding entry in paper.bib
