### 2026-03-25 jchejarla/spring-batch-db-cluster-partitioning#10 @jchejarla
Both items addressed:

Automated test pipeline: The GitHub Actions CI workflow (.github/workflows/maven.yml) has been enhanced to:

1) Run mvn clean verify (builds + executes all unit tests) on every push and pull request
2) Test against a Java 17 + Java 21 matrix
3) Upload test reports as build artifacts for inspection
4) The repository already has 20+ unit tests covering partitioning strategies, cluster node management, partition assignment, aggregation, auto-configuration, and polling.

CONTRIBUTING.md: Completely rewritten with detailed sections for:

1)Reporting bugs (what to include: steps to reproduce, environment details, stack traces)
2) Suggesting features
3) Development setup (prerequisites, build commands, project structure)
4) Submitting a pull request (fork -> branch -> code -> test -> commit -> PR workflow)
5) Code style guidelines
6) Testing guidelines (unit test vs integration test conventions)
