### 2026-03-25 jchejarla/spring-batch-db-cluster-partitioning#11 @jchejarla
Thank you for raising this. A new Project Maturity section has been added to the paper providing context on:

1) Publication to Maven Central since v1.0.0, now at v2.0.0
2) v2.0.0 additions: scale-up partition strategy, actuator health endpoints, multi-database support (PostgreSQL, MySQL, Oracle, H2)
3) Comprehensive unit and integration test suite executed via GitHub Actions CI on every push/PR
4) A runnable examples/ module with Docker-based setup for local validation
5) Documentation, issue templates, and a contributor guide maintained in the repository
6) The author is committed to long-term maintenance and actively encourages community contributions through the issue tracker and pull requests
