### 2026-03-25 jchejarla/spring-batch-db-cluster-partitioning#12 @jchejarla
Addressed. The Statement of Need now includes a concrete performance paragraph:

"The performance benefit is direct: a job that takes 60 minutes on a single node can complete in approximately 15 minutes across four workers using round-robin distribution, with near-linear speedup for I/O-bound workloads. The extension adds only lightweight SQL operations (heartbeat writes, partition claim updates) whose overhead is negligible relative to the batch workload itself."

Additionally, the new **Software Design** section provides significantly more technical depth — covering the coordination table schema, the full partitioning flow, distribution strategies, the two-phase failure detection protocol, and the transferable vs non-transferable partition recovery policy.
