### 2026-03-25 252f796
to use maven-gpg-plugin plugin only when releasing a new version of the jar
