### 2026-03-25 #9460 @jchejarla
Hi @david-guzman @Hardik27  - thank you for the feedback. I've addressed all the issues raised and closed them in the repo:
Here is the summary of the changes

Paper (paper.md) — major rewrite:

1) Summary expanded for a non-specialist audience
2) Statement of Need now includes research domains (bioinformatics, finance, climate), target audience, why Spring-centric vs Jakarta EE, and concrete performance benefits
3) Software Design section added — coordination tables, partitioning flow, two-phase failure detection, architecture diagram
4) Research Impact section added
5) AI Usage Disclosure section added
6) All sections rewritten in prose (no bullet-point lists); consistent "extension" naming; "small-to-medium" quantified as ~2–20 nodes; diagram renders as PNG
7) References cleaned up — all use proper citation syntax with matching bib entries
8) Word count: 614 -> ~1,550

Repository improvements:

1) examples/README.md now includes example job descriptions, quick-start guide, failover testing steps, monitoring endpoints, project structure, and configuration reference
2) CONTRIBUTING.md rewritten with detailed sections for bug reporting, development setup, build commands, PR workflow, code style, and testing guidelines
3) CI pipeline enhanced with Java 17+21 matrix, explicit test execution, and test report uploads
4) Project Maturity section in paper addresses community engagement context
5) All changes are on main. Could you please regenerate the paper proof when ready?
