### 2026-03-30 db0fa5e
update to spring boot 4, update tests, align junit version
