+++ examples/src/main/java/examples/io/github/jchejarla/springbatch/clustering/common/RangeItemReader.java
@@ -1,9 +1,10 @@
 package examples.io.github.jchejarla.springbatch.clustering.common;
 
-import org.springframework.batch.item.ItemReader;
-import org.springframework.batch.item.NonTransientResourceException;
-import org.springframework.batch.item.ParseException;
-import org.springframework.batch.item.UnexpectedInputException;
+
+import org.springframework.batch.infrastructure.item.ItemReader;
+import org.springframework.batch.infrastructure.item.NonTransientResourceException;
+import org.springframework.batch.infrastructure.item.ParseException;
+import org.springframework.batch.infrastructure.item.UnexpectedInputException;
 
 public class RangeItemReader implements ItemReader<Range> {
 
