+++ examples/src/main/java/examples/io/github/jchejarla/springbatch/clustering/RestEndPoints.java
@@ -3,10 +3,10 @@ package examples.io.github.jchejarla.springbatch.clustering;
 import examples.io.github.jchejarla.springbatch.clustering.simplejob.SimpleJobConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobParameters;
-import org.springframework.batch.core.JobParametersBuilder;
+import org.springframework.batch.core.job.Job;
+import org.springframework.batch.core.job.JobExecution;
+import org.springframework.batch.core.job.parameters.JobParameters;
+import org.springframework.batch.core.job.parameters.JobParametersBuilder;
 import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.context.ApplicationContext;
 import org.springframework.http.ResponseEntity;
@@ -42,7 +42,7 @@ public class RestEndPoints {
             long startTime = System.currentTimeMillis();
             JobExecution jobExecution = jobLauncher.run(job, parameters);
             long endTime = System.currentTimeMillis();
-            String sb = "Job Id : " + jobExecution.getJobId() + " Completed in " + (endTime - startTime) + " milli seconds." + "\n" +
+            String sb = "Job Id : " + jobExecution.getJobInstanceId() + " Completed in " + (endTime - startTime) + " milli seconds." + "\n" +
                     "Output: " + "converted "+rows +" rows CSV data into XML format";
             return ResponseEntity.ok(sb);
         } catch(Exception e) {
@@ -65,7 +65,7 @@ public class RestEndPoints {
             JobExecution jobExecution = jobLauncher.run(job, parameters);
             long endTime = System.currentTimeMillis();
             SimpleJobConfig simpleJobConfig = applicationContext.getBean(SimpleJobConfig.class);
-            String sb = "Job Id : " + jobExecution.getJobId() + " Completed in " + (endTime - startTime) + " milli seconds." + "\n" +
+            String sb = "Job Id : " + jobExecution.getJobInstanceId() + " Completed in " + (endTime - startTime) + " milli seconds." + "\n" +
                     "Output: " + "sum of number from " + from + " to " + to + " is "+
                     simpleJobConfig.getSumAggregatorCallback().getSum();
             return ResponseEntity.ok(sb);
@@ -89,7 +89,7 @@ public class RestEndPoints {
             JobExecution jobExecution = jobLauncher.run(job, parameters);
             Long result = (Long) jobExecution.getExecutionContext().get("totalSum");
             long endTime = System.currentTimeMillis();
-            String sb = "Job Id : " + jobExecution.getJobId() + " Completed in " + (endTime - startTime) + " milli seconds." + "\n" +
+            String sb = "Job Id : " + jobExecution.getJobInstanceId() + " Completed in " + (endTime - startTime) + " milli seconds." + "\n" +
                     "Output: " + "sum of number from " + from + " to " + to + " is " +result;
             return ResponseEntity.ok(sb);
         } catch(Exception e) {
@@ -112,7 +112,7 @@ public class RestEndPoints {
             JobExecution jobExecution = jobLauncher.run(job, parameters);
             Long result = (Long) jobExecution.getExecutionContext().get("totalSum");
             long endTime = System.currentTimeMillis();
-            String sb = "Job Id : " + jobExecution.getJobId() + " Completed in " + (endTime - startTime) + " milli seconds." + "\n" +
+            String sb = "Job Id : " + jobExecution.getJobInstanceId() + " Completed in " + (endTime - startTime) + " milli seconds." + "\n" +
                     "Output: " + "sum of number from " + from + " to " + to + " is " +result;
             return ResponseEntity.ok(sb);
         } catch(Exception e) {
