+++ spring-batch-db-cluster-core/src/main/java/io/github/jchejarla/springbatch/clustering/actuate/BatchClusterNodeHealthIndicator.java
@@ -6,9 +6,9 @@ import io.github.jchejarla.springbatch.clustering.mgmt.ClusterNodeManager;
 import io.github.jchejarla.springbatch.clustering.mgmt.NodeLoad;
 import io.github.jchejarla.springbatch.clustering.mgmt.NodeStatus;
 import lombok.RequiredArgsConstructor;
-import org.springframework.boot.actuate.autoconfigure.health.ConditionalOnEnabledHealthIndicator;
-import org.springframework.boot.actuate.health.Health;
-import org.springframework.boot.actuate.health.HealthIndicator;
+import org.springframework.boot.health.autoconfigure.contributor.ConditionalOnEnabledHealthIndicator;
+import org.springframework.boot.health.contributor.Health;
+import org.springframework.boot.health.contributor.HealthIndicator;
 
 import java.util.HashMap;
 import java.util.Map;
