### 2026-05-04 #9460 @majensen
@jchejarla I've added myself as a reviewer, since I walked through the installation and examples myself, and @Hardik27 did not create a checklist, although the issues he raised indicate what he thought was missing.

The paper is well-written overall- I will follow up with a few minor edits here, hopefully in the next couple of days.

Can you create a persistent archive of the codebase using Zenodo as in [this example](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729)? Please use the name of the paper for the archive name. Just provide me back here in the review with the DOI that is assigned. Thanks!
