### 2026-05-10 #9460 @danielskatz
👋 @majensen - I'm a bit confused here, as I only see one complete review: yours. I also see two partially-complete reviews from @Hardik27 & @david-guzman.  Can you help me understand what has happened here, and how the minimum of two complete reviews that JOSS requires has been met?
