### 2026-05-10 8ab626a
Fix reference capitalization per JOSS reviewer feedback
