+++ paper.bib
@@ -40,7 +40,7 @@
 }
 
 @article{ludascher2006kepler,
-  title   = {Scientific Workflow Management and the Kepler System},
+  title   = {Scientific Workflow Management and the {Kepler} System},
   author  = {Lud{\"a}scher, Bertram and Altintas, Ilkay and Berkley, Chad and Higgins, Dan and Jaeger, Efrat and Jones, Matthew and Lee, Edward A. and Tao, Jing and Zhao, Yang},
   journal = {Concurrency and Computation: Practice and Experience},
   volume  = {18},
@@ -51,7 +51,7 @@
 }
 
 @article{wolstencroft2013taverna,
-  title   = {The Taverna Workflow Suite: Designing and Executing Workflows of Web Services on the Desktop, Web or in the Cloud},
+  title   = {The {Taverna} Workflow Suite: Designing and Executing Workflows of Web Services on the Desktop, Web or in the Cloud},
   author  = {Wolstencroft, Katherine and Haines, Robert and Fellows, Duncan and Williams, Alan and Withers, David and Owen, Stuart and Soiland-Reyes, Stian and Dunlop, Ian and Nenadic, Aleksandra and Fisher, Paul and Bhagat, Jiten and Belhajjame, Khalid and Hardisty, Alex and de la Hidalga, Abraham Nieva and Vargas, Marco Roos and Goble, Carole},
   journal = {Nucleic Acids Research},
   volume  = {41},
@@ -62,7 +62,7 @@
 }
 
 @book{tanenbaum2007distributed,
-  title   = {Distributed Systems: Principles and Paradigms},
+  title   = {{Distributed} {Systems}: {Principles} and {Paradigms}},
   author  = {Tanenbaum, Andrew S. and Van Steen, Maarten},
   edition = {2nd},
   publisher = {Pearson Prentice Hall},
@@ -79,7 +79,7 @@
 }
 
 @inproceedings{burrows2006chubby,
-  title   = {The Chubby Lock Service for Loosely-Coupled Distributed Systems},
+  title   = {The {Chubby} Lock Service for Loosely-Coupled Distributed Systems},
   author  = {Burrows, Michael},
   booktitle = {Proceedings of the 7th USENIX Symposium on Operating Systems Design and Implementation (OSDI)},
   year    = {2006},
@@ -87,7 +87,7 @@
 }
 
 @inproceedings{ongaro2014raft,
-  title   = {In Search of an Understandable Consensus Algorithm (Raft)},
+  title   = {In Search of an Understandable Consensus Algorithm ({Raft})},
   author  = {Ongaro, Diego and Ousterhout, John},
   booktitle = {USENIX Annual Technical Conference (ATC)},
   year    = {2014},
